test: move packages
